+++ desktop/package.json
@@ -3,15 +3,6 @@
   "private": true,
   "dependencies": {
     "@foxglove/studio-base": "workspace:packages/studio-base",
-    "@foxglove/studio-firebase": "workspace:packages/studio-firebase",
-    "jest": "27.0.0-next.8"
-  },
-  "devDependencies": {
-    "@types/plist": "^3",
-    "plist": "3.0.2",
-    "quicklookjs": "0.0.2"
-  },
-  "scripts": {
-    "test": "jest"
+    "@foxglove/studio-firebase": "workspace:packages/studio-firebase"
   }
 }
